+++ src/AppBundle/Form/Type/Text/TitleType.php
@@ -1,12 +1,12 @@
 <?php
 
-namespace AppBundle\Form\Type;
+namespace AppBundle\Form\Type\Text;
 
 use Symfony\Component\Form\AbstractType;
 use Symfony\Component\Form\FormBuilderInterface;
 use Symfony\Component\OptionsResolver\OptionsResolver;
 
-class TextType extends AbstractType
+class TitleType extends AbstractType
 {
     /**
      * {@inheritdoc}
